:sparkles: update UI seekbar in item_monthly_payment_main.xml.kt :sparkles: update int to float in LoanCalculatorState.kt :sparkles: update in to float in LoansCalculatorViewModel.kt :sparkles: update int to float MonthlyPaymentBreakdownMainModel.kt :lipstick: update UI in non non citizen screen

+++ app/src/main/java/com/unionbankph/corporate/loan/calculator/LoanCalculatorState.kt
@@ -5,10 +5,10 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class LoanCalculatorState (
-    var amount: Int? = null,
-    var principal: Int? = null,
-    var interest: Int? = null,
+    var amount: Float? = null,
+    var principal: Float? = null,
+    var interest: Float? = null,
     var month: Int? = null,
-    var monthlyPayment: Int? = null,
-    var totalAmountPayable: Int? = null
+    var monthlyPayment: Float? = null,
+    var totalAmountPayable: Float? = null
 )
