+++ app/src/main/java/com/unionbankph/corporate/loan/calculator/LoansCalculatorCallback.kt
@@ -2,11 +2,11 @@ package com.unionbankph.corporate.loan.calculator
 
 interface LoansCalculatorCallback {
 
-    fun onAmountChange(amount: Int)
+    fun onAmountChange(amount: Float)
     fun onMonthsChange(months: Int)
-    fun onTotalAmountPayable(totalAmountPayable: Int)
-    fun onMonthlyPayment(monthlyPayment: Int)
-    fun onPrincipal(principal: Int)
-    fun onInterest(interest: Int)
+    fun onTotalAmountPayable(totalAmountPayable: Float)
+    fun onMonthlyPayment(monthlyPayment: Float)
+    fun onPrincipal(principal: Float)
+    fun onInterest(interest: Float)
     fun onApplyNow()
 }
